+++ Popup/GoalPREditPopup.xaml.cs
@@ -136,7 +136,7 @@ public partial class GoalPREditPopup
             {
                 weight_update_string = weight_update_string.ToString();
 
-                int weight_update = int.Parse(weight_update_string);
+                double weight_update = double.Parse(weight_update_string);
                 await App.RecordRepo.Edit_Goal_PR(name, date, has_desired, true, weight_update, -1, -1, -1);
 
                 error_prompt.IsVisible = false;
